Do not require X- prefix for custom headers
RFC 6648, released in June 2012, deprecated the practice of prefixing the names of unstandardized parameters with "X-" in all IETF protocols. In particular, section 3 recommends that "creators of new parameters to be used in the context of application protocols ... SHOULD NOT prefix their parameter names with 'X-' or similar constructs."

Since the IETF discourages HTTP protocol developers from prefixing their header names with "X-", the Sphinx HTTP domain should no longer encourage them to do the same. So, this commit eliminates warnings for unrecognized HTTP headers, even the ones that lack an "X-" prefix.

 One might consider going farther and issuing warnings for header names with the "X-" prefix, but that question is left unresolved here.